requirements.txt
aioredis==2.0.1
anyio==3.6.1
asgiref==3.5.2
async-timeout==4.0.2
click==8.1.3
colorama==0.4.4
fastapi==0.78.0
h11==0.13.0
idna==3.3
iso8601==1.0.2
prometheus-client==0.14.1
PyYAML==6.0
sniffio==1.2.0
starlette==0.19.1
starlette-exporter==0.13.0
typing_extensions==4.2.0
ujson==5.3.0
uvicorn==0.17.6
pottery
redis
celery
python-arango
pytz
pyhumps
pydantic[email]
loguru
python-multipart
pytest-asyncio
onetimepad
dictdiffer
aioarango